pair rulesets explicitly to some specific transit data format
This commit introduces transit data format as part of core data model and adds checks and data generation limitations based on its applicability. This will allow users to still request arbitrary formats, but now we will be more strict about whether we'll do any processing based on them or not.